Fixed small bug when selecting an ingredient and fixed a bug with the restart/food action button
Before the user was able to pick 1 ingredient multiple times and it would end the game but now its fixed. the user will only be able to select one ingredient.

When i would restart the game the food action button would stay on on the screen.now this problem is fixed